+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Improved action and gutter icons to work with high contrast theme
 - Use non-null values in editor state
 - Migrate to Kotlin UI DSL version 2
+- Move settings icon to the left toolbar in 3D preview
 - Update GitHub Actions build pipeline
 - Update platform version to `2021.3.3`
 - `pluginVerifierIdeVersions` – upgrade to `2021.3.3, 2022.1`